发布版本V1.0.0

+++ src/main/kotlin/moe/rainbowyang/math/Math.kt
@@ -4,8 +4,8 @@ import kotlin.math.*
 
 fun lengthOf(vararg nums: Number) = sqrt(sumOfSquare(*nums))
 fun <N : Number> List<N>.lengthOf() = sqrt(sumOfSquare())
-fun sumOfSquare(vararg nums: Number) = nums.sumByDouble { it.toDouble() * it.toDouble() }
-fun <N : Number> List<N>.sumOfSquare() = sumByDouble { it.toDouble() * it.toDouble() }
+fun sumOfSquare(vararg nums: Number) = nums.sumByDouble { it * it }
+fun <N : Number> List<N>.sumOfSquare() = sumByDouble { it * it }
 fun checkValues(vararg values: Number) = values.all { !it.toDouble().isNaN() || !it.toDouble().isInfinite() }
 fun <N : Number> List<N>.checkValues() = all { !it.toDouble().isNaN() || !it.toDouble().isInfinite() }
 
@@ -18,4 +18,4 @@ fun until(end: Int) = 0 until end
 
 infix fun Number.almostEquals(other: Number) = this.almostEquals(other, 10)
 fun Number.almostEquals(other: Number, accuracy: Int) =
-        abs(this.toDouble() - other.toDouble()) < (0.1.pow(accuracy.toDouble()))
+        abs(this - other) < (0.1.pow(accuracy.toDouble()))
